LayoutTests/imported/w3c/web-platform-tests/css/css-highlight-api/HighlightRegistry-maplike-tampered-Map-prototype-expected.txt


PASS HighlightRegistry is a maplike interface that works as expected even if Map.prototype is tampered.
